requirements.txt
streamlit
google-cloud-firestore
google-cloud-secret-manager
streamlit-firebase-auth # または選定したライブラリ名